（1）新增从数据库读入路网的模块；（2）适配路网新的数据结构：2.1对象id由int改为long；2.2车道线由多段折线构成；2.3顺接车道连接器；2.4路段末端车道转向信息（U/S/L/R）

+++ TranSysLab/src/main/java/com/transyslab/roadnetwork/GeoSurface.java
@@ -9,18 +9,18 @@ public class GeoSurface implements NetworkObject{
 	protected List<GeoPoint> kerbPoints;
 	protected int segmentId;
 	protected int index;
-	protected int id;
+	protected long id;
 	protected String objInfo;
 	protected boolean isSelected;
 	public GeoSurface(){
 		kerbPoints = new ArrayList<GeoPoint>();
 	}
-	public void init(int code, int segmentid) {
+	public void init(long code, int segmentid) {
 		this.id =  code;
 		segmentId = segmentid;
 
 	}
-	public int getId(){
+	public long getId(){
 		return this.id;
 	}
 	public void setSelected(boolean flag){
